requirements.txt
streamlit==1.43.2
nlpaug==1.1.11
beautifulsoup4==4.13.3
transformers==4.50.0
nltk==3.9.1
spacy==3.8.3
contractions==0.1.73
spellchecker==0.4
pyspellchecker==0.8.2
matplotlib==3.9.4
scikit_learn==1.6.1
Requests==2.32.3
gensim==4.3.3 
numpy==1.26.4
PyPDF2==3.0.1
lxml==5.4.0
python-docx==1.1.2